More commitment to unsigned limit values
A lot of values that should have been unsigned ints were ints because of how C# is "allergic" to unsigned indexing in lists. This causes caused confusion on whether signed or unsigned ints should be used in certain sitatutions which caused odd mixings. While probably not 100% fixed, many signed ints have been commited to unsigned it.