Refactored segments for elem and data
The default data for Memories (Data) and Tables (Elem) have been refactored and semi-unified.
The both use the same base DefSegment struct.
Multiple Data and Elem statements are now allowed.
Offsets for Data and Elem are handled in a unified way.

I currently don't know the extent of where globals are allowed in expressions, but the code is also refactored better handle them - although changes are more understanding and a future refactor will be needed.